backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
yfinance==0.2.51
pandas==2.2.3
numpy==2.2.0
scipy==1.15.0
newsapi-python==0.2.7
python-dotenv==1.0.1
httpx==0.28.1
